Moved interfaces out of unnecessary Contracts project into Core
