+++ CareTogetherCMS.sln
@@ -40,8 +40,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Prod", "Prod", "{2DC7B8F3-2
 		src\b2c\CustomPolicies\Prod\TrustFrameworkExtensions.xml = src\b2c\CustomPolicies\Prod\TrustFrameworkExtensions.xml
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CareTogether.Contracts", "src\CareTogether.Contracts\CareTogether.Contracts.csproj", "{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CareTogether.Core", "src\CareTogether.Core\CareTogether.Core.csproj", "{82B69D6F-DA25-452B-883D-71B9AE91DF23}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CareTogether.Core.Test", "test\CareTogether.Core.Test\CareTogether.Core.Test.csproj", "{B2CA1F10-216E-404E-A3C2-70423D5BCF5D}"
@@ -67,10 +65,6 @@ Global
 		{67F0E369-C927-4278-A7D0-DDFBE029003F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67F0E369-C927-4278-A7D0-DDFBE029003F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67F0E369-C927-4278-A7D0-DDFBE029003F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{82B69D6F-DA25-452B-883D-71B9AE91DF23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{82B69D6F-DA25-452B-883D-71B9AE91DF23}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{82B69D6F-DA25-452B-883D-71B9AE91DF23}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -93,7 +87,6 @@ Global
 		{E477275E-8530-49B4-ADBB-D61DD5E79BD6} = {7787CCCA-D8C1-4E53-8EBA-1A07B5DEA5AB}
 		{7B8B0240-E608-4349-89AD-AB5E410BD273} = {E477275E-8530-49B4-ADBB-D61DD5E79BD6}
 		{2DC7B8F3-2B29-4EC9-BFFF-CC49C817469C} = {E477275E-8530-49B4-ADBB-D61DD5E79BD6}
-		{0BEB306D-AA4B-4BB9-A8CC-E5632B63F3E7} = {465BA74E-CEF3-4608-8809-8C7A2E979DAC}
 		{82B69D6F-DA25-452B-883D-71B9AE91DF23} = {465BA74E-CEF3-4608-8809-8C7A2E979DAC}
 		{B2CA1F10-216E-404E-A3C2-70423D5BCF5D} = {42D43EAF-EE89-479A-B304-3B5DF56A3494}
 		{7DC7327F-6F01-4561-943D-D385FB5637E1} = {42D43EAF-EE89-479A-B304-3B5DF56A3494}
