--- src/CareTogether.Contracts/CareTogether.Contracts.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>CareTogether</RootNamespace>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="JsonPolymorph" Version="1.0.1" />
-  </ItemGroup>
-  
-</Project>
