--- src/CareTogether.Contracts/Managers/IApprovalManager.cs
@@ -1,16 +1,0 @@
-using CareTogether.Resources;
-using System;
-using System.Security.Claims;
-using System.Threading.Tasks;
-
-namespace CareTogether.Managers
-{
-    public interface IApprovalManager
-    {
-        Task<CombinedFamilyInfo> ExecuteVolunteerFamilyCommandAsync(Guid organizationId, Guid locationId,
-            ClaimsPrincipal user, VolunteerFamilyCommand command);
-
-        Task<CombinedFamilyInfo> ExecuteVolunteerCommandAsync(Guid organizationId, Guid locationId,
-            ClaimsPrincipal user, VolunteerCommand command);
-    }
-}
