--- src/CareTogether.Contracts/Managers/IReferralManager.cs
@@ -1,16 +1,0 @@
-﻿using CareTogether.Resources;
-using System;
-using System.Security.Claims;
-using System.Threading.Tasks;
-
-namespace CareTogether.Managers
-{
-    public interface IReferralManager
-    {
-        Task<CombinedFamilyInfo> ExecuteReferralCommandAsync(Guid organizationId, Guid locationId,
-            ClaimsPrincipal user, ReferralCommand command);
-
-        Task<CombinedFamilyInfo> ExecuteArrangementCommandAsync(Guid organizationId, Guid locationId,
-            ClaimsPrincipal user, ArrangementCommand command);
-    }
-}
