--- src/CareTogether.Contracts/Resources/Storage/IFileStore.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace CareTogether.Resources.Storage
-{
-    public interface IFileStore
-    {
-        Task<Uri> GetValetReadUrlAsync(Guid organizationId, Guid locationId, Guid documentId);
-
-        Task<Uri> GetValetCreateUrlAsync(Guid organizationId, Guid locationId, Guid documentId);
-    }
-}
