+++ src/CareTogether.Core/Resources/Storage/IFileStore.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace CareTogether.Resources.Storage
+{
+    public interface IFileStore
+    {
+        Task<Uri> GetValetReadUrlAsync(Guid organizationId, Guid locationId, Guid documentId);
+
+        Task<Uri> GetValetCreateUrlAsync(Guid organizationId, Guid locationId, Guid documentId);
+    }
+}
